Dataflow/pipeline/requirements.txt
apache-beam==2.34.0
avro-python3==1.9.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.7
fasteners==0.16.3
future==0.18.2
google-api-core==2.2.2
google-apitools==0.5.32
google-auth==2.3.3
google-cloud-bigquery==2.30.1
google-cloud-bigquery-storage==2.10.1
google-cloud-bigtable==2.4.0
google-cloud-core==2.2.1
google-cloud-datastore==2.4.0
google-cloud-dlp==3.3.1
google-cloud-language==2.3.1
google-cloud-pubsub==2.9.0
google-cloud-recommendations-ai==0.5.1
google-cloud-spanner==3.11.1
google-cloud-storage==1.42.3
google-cloud-videointelligence==2.5.1
google-cloud-vision==2.6.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-gcp==0.2.2
grpcio-status==1.41.1
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==4.1.3
orjson==3.6.4
packaging==21.2
proto-plus==1.19.8
protobuf==3.19.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7